Update TaxCalculator.java
while loop so that it keeps asking until the user wants to exit and if exit it returns the main. I have to implement a validation or like a exception handling. There is also a issue that it can lead to negative numbers of tax which needs to be fixed.