Update TemperatureConverter.java
Similar to the length converter added while true loop so user can make multiple calculations and has the option to continue or go back to the main. There is an inconsistency in the way of asking the user to leave or continue across the project which needs to be fixed. Also needs a validation.